+++ Frends.Smb.CreateDirectory/Frends.Smb.CreateDirectory/Definitions/AuthenticationMode.cs
@@ -14,4 +14,10 @@ public enum AuthenticationMode
     /// Kerberos authentication protocol.
     /// </summary>
     Kerberos,
+
+    /// <summary>
+    /// Kerberos authentication protocol using a pre-existing ticket cache (ccache) file.
+    /// Requires a valid TGT in the ccache file created externally e.g. via kinit.
+    /// </summary>
+    KerberosTicketCache,
 }